Improve Markdown by handling some unconverted entities
Markdown apparently tries to figure out when the input already contains tags and
entities, and leaves them alone if found. In this case the input never contains
tags or entities, so that caution is unwarranted and results in invalid entities.
This patch handles those situations explicitly, plus a little formatting cleanup
and so forth.